requirements.txt
pandas>=2.0.0
pathlib
numpy
python-dotenv
